requirements.txt
# Core dependencies
torch>=2.0.0
torch-geometric>=2.3.0
rdkit-pypi>=2022.9.5
numpy>=1.21.0
scipy>=1.7.0

# Data processing
pandas>=1.3.0
h5py>=3.7.0

# Configuration and utilities
pyyaml>=6.0
omegaconf>=2.3.0
hydra-core>=1.3.0

# Visualization and plotting
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Machine learning utilities
scikit-learn>=1.0.0
tqdm>=4.64.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Logging and experiment tracking
wandb>=0.13.0
tensorboard>=2.10.0

# Optional: For CUDA support (uncomment if needed)
# torch-scatter>=2.1.0
# torch-sparse>=0.6.0
# torch-cluster>=1.6.0